+++ src/oihana/enums/Order.php
@@ -9,12 +9,22 @@ class Order
     use ConstantsTrait ;
 
     /**
-     * The ascending order.
+     * The ascending order (lower case).
+     */
+    public const string asc  = 'asc' ;
+
+    /**
+     * The ascending order (upper case).
      */
     public const string ASC  = 'ASC' ;
 
     /**
-     * The descending order.
+     * The descending order (lower case).
+     */
+    public const string desc = 'desc' ;
+
+    /**
+     * The descending order (upper case).
      */
     public const string DESC = 'DESC' ;
 }