+++ CHANGELOG.md
@@ -14,10 +14,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Adds oihana\files\assertWritableDirectory
 - Adds oihana\files\createDirectory
 - Adds oihana\files\deleteDirectory
+- Adds oihana\files\findFiles
 - Adds oihana\files\getBaseFileName
 - Adds oihana\files\getDirectory
 - Adds oihana\files\getFileExtension
-- Adds oihana\files\listFiles
+- Adds oihana\files\makeTimestampedDirectory
 - Adds oihana\files\sortFiles
 - Adds oihana\files\enums\FileExtension
 - Adds oihana\files\enums\FileMimeType